backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
psycopg2==2.9.10
django-cors-headers==4.7.0
drf-yasg==1.21.10
python-decouple==3.8
pyjwt==2.8.0
supabase==2.15.1
underthesea==6.8.4
scikit-learn==1.7.1
scipy==1.16.1
joblib==1.5.2
django-celery-beat==2.8.1
redis==5.3.1